refactor(mcp): extract exit_events.ts for server-only exit/pause event processing
Extract getExitOrPauseEvents to a separate exit_events.ts file to isolate
server-only code that depends on resources.
